better folder structure for core module
