+++ src/app/api/v1/rate_limits.py
@@ -6,7 +6,7 @@ import fastapi
 
 from app.api.dependencies import get_current_superuser
 from app.api.paginated import PaginatedListResponse, paginated_response, compute_offset
-from app.core.database import async_get_db
+from src.app.core.db.database import async_get_db
 from app.crud.crud_rate_limit import crud_rate_limits
 from app.crud.crud_tier import crud_tiers
 from app.schemas.rate_limit import (
