+++ src/app/api/v1/tasks.py
@@ -1,7 +1,7 @@
 from arq.jobs import Job as ArqJob
 from fastapi import APIRouter, Depends
 
-from app.core import queue
+from src.app.core.utils import queue
 from app.schemas.job import Job
 from app.api.dependencies import rate_limiter
