+++ src/app/core/db/crud_token_blacklist.py
@@ -1,6 +1,6 @@
 from app.crud.crud_base import CRUDBase
-from app.models.token_blacklist import TokenBlacklist
-from app.schemas.token_blacklist import TokenBlacklistCreate, TokenBlacklistUpdate
+from src.app.core.db.token_blacklist import TokenBlacklist
+from src.app.core.schemas import TokenBlacklistCreate, TokenBlacklistUpdate
 
 CRUDTokenBlacklist = CRUDBase[TokenBlacklist, TokenBlacklistCreate, TokenBlacklistUpdate, TokenBlacklistUpdate, None]
 crud_token_blacklist = CRUDTokenBlacklist(TokenBlacklist)
