+++ src/app/core/exceptions/exceptions.py
