+++ src/app/core/schemas.py
@@ -1,0 +1,60 @@
+import uuid as uuid_pkg
+from datetime import datetime
+
+from pydantic import BaseModel, Field, field_serializer
+
+class HealthCheck(BaseModel):
+    name: str
+    version: str
+    description: str
+
+# -------------- mixins --------------
+class UUIDSchema(BaseModel):
+    uuid: uuid_pkg.UUID = Field(default_factory=uuid_pkg.uuid4)
+
+
+class TimestampSchema(BaseModel):
+    created_at: datetime = Field(default_factory=datetime.utcnow)
+    updated_at: datetime = Field(default=None)
+
+    @field_serializer("created_at")
+    def serialize_dt(self, created_at: datetime | None, _info):
+        return created_at.isoformat()
+
+    @field_serializer("updated_at")
+    def serialize_updated_at(self, updated_at: datetime | None, _info):
+        if updated_at is not None:
+            return updated_at.isoformat()
+
+
+class PersistentDeletion(BaseModel):
+    deleted_at: datetime | None = Field(default=None)
+    is_deleted: bool = False
+
+    @field_serializer('deleted_at')
+    def serialize_dates(self, deleted_at: datetime | None, _info):
+        if deleted_at is not None:
+            return deleted_at.isoformat()
+
+
+# -------------- token --------------
+class Token(BaseModel):
+    access_token: str
+    token_type: str
+
+
+class TokenData(BaseModel):
+    username_or_email: str
+
+
+class TokenBlacklistBase(BaseModel):
+    token: str
+    expires_at: datetime
+
+
+class TokenBlacklistCreate(TokenBlacklistBase):
+    pass
+
+
+class TokenBlacklistUpdate(TokenBlacklistBase):
+    pass
