+++ src/app/core/utils/cache.py
@@ -7,7 +7,7 @@ from fastapi import Request, Response
 from fastapi.encoders import jsonable_encoder
 from redis.asyncio import Redis, ConnectionPool
 
-from app.core.exceptions import CacheIdentificationInferenceError, InvalidRequestError
+from src.app.core.exceptions.exceptions import CacheIdentificationInferenceError, InvalidRequestError
 
 pool: ConnectionPool | None = None
 client: Redis | None = None
