+++ src/app/models/tier.py
@@ -4,7 +4,7 @@ from datetime import datetime
 from sqlalchemy import String, DateTime
 from sqlalchemy.orm import Mapped, mapped_column
 
-from app.core.database import Base
+from src.app.core.db.database import Base
 
 class Tier(Base):
     __tablename__ = "tier"
