+++ src/app/models/user.py
@@ -5,7 +5,7 @@ from datetime import datetime
 from sqlalchemy import String, DateTime, ForeignKey
 from sqlalchemy.orm import Mapped, mapped_column
 
-from app.core.database import Base
+from src.app.core.db.database import Base
 
 class User(Base):
     __tablename__ = "user"
