--- src/app/schemas/token_blacklist.py
@@ -1,14 +1,0 @@
-from pydantic import BaseModel
-from datetime import datetime
-
-class TokenBlacklistBase(BaseModel):
-    token: str
-    expires_at: datetime
-
-
-class TokenBlacklistCreate(TokenBlacklistBase):
-    pass
-
-
-class TokenBlacklistUpdate(TokenBlacklistBase):
-    pass
