+++ src/scripts/create_first_superuser.py
@@ -11,12 +11,12 @@ from sqlalchemy import (
     DateTime, 
     Boolean
 )
-from app.core.database import async_engine
+from src.app.core.db.database import async_engine
 from sqlalchemy.dialects.postgresql import UUID
 import uuid
 from datetime import datetime
 
-from app.core.database import AsyncSession, local_session
+from src.app.core.db.database import AsyncSession, local_session
 from app.core.config import settings
 from app.models.user import User
 from app.core.security import get_password_hash
